remove qcom.hw.aac.encoder
mako doesn't support this because they have QC enhancements disabled.
But since we need this flag to fix lags it's used on this device.
The problem is that the aac encoder doesn't seem to work.
It crashes with a message which tells you that the audio format
'audio/raw' isn't supported

Change-Id: Ic6a8221af9a7ebe1be263b3d0b2d89653c5755a8