Aries: fix 'Battery Saver' profile
Switched from conservative to powersave as governor for 'Battery Saver'
profile because conservative is not supported.
This cause that when you switch profiles using the toggle, before the 'Battery Saver' profile, there's 'Performances' that sets governor as performances
and then you switch to 'Battery Saver' but governor is still performances
and this result in a battery drain for who want to save battery using these profiles.

Change-Id: Ieb81b5dc5554944fb23f2f1c9baad7bfee2aa7dd